backend/requirements.txt
fastapi
uvicorn
joblib
numpy
python-dotenv
google-generativeai
